Applied basic python packagin structure and modularization of some classes.
