+++ differential_evolution/__init__.py
@@ -1,0 +1,2 @@
+from . import algorithm
+from . import examples
