+++ differential_evolution/algorithm/__init__.py
@@ -1,0 +1,1 @@
+from .differential_evolution import DifferentialEvolution
