+++ differential_evolution/examples/__init__.py
@@ -1,0 +1,2 @@
+from .simple_toy import simple_toy
+from .full_toy import full_toy
